refactor: extract WeeklySnapshotCard with WeeklySnapshot.compute()
Snapshot computation becomes a static method with a unit test;
sparkline subview travels with it. Card takes an onTap closure
so JournalView can wire it differently from TodayView.
